Update I Added a `managers` file
you can keep your models lean and clean and separate the model logic from manager-level functionality
